Choose not to support SSL.
We want to keep this library as small in scope as possible,
optimized for use on constrained platforms. Supporting SSL
would complicate deployment of user apps using the library,
as well as increasing the distribution size.

If there is a demand for SSL support in the future, we can
try to make it an opt-in feature that won't burden other users.